[UI] Исправлено отображение пользователя без аватара Resolve #1207
